pyproject.toml
[project]
name = "agentic-cookbook"
version = "0.1.0"
description = "A barebones implementation of agentic architectures for educational and prototyping purposes."
authors = [
    {name = "Your Name", email = "you@example.com"}
]
dependencies = [
    "litellm>=1.40.0",
    "rich>=13.7.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.7.0",
]
requires-python = ">=3.10"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "tools*"]
